refactor(architecture): move testing packages next to their corresponding modules
