+++ antenna-basic-assembly/antenna-cli/pom.xml
@@ -126,7 +126,7 @@
         <!-- ################################ testing dependencies ################################ -->
         <dependency>
             <groupId>org.eclipse.sw360.antenna</groupId>
-            <artifactId>antenna-frontend-stubs-testing</artifactId>
+            <artifactId>frontend-stubs-testing</artifactId>
             <version>${project.version}</version>
             <scope>test</scope>
         </dependency>
