+++ antenna/abstract-compliance-checker/pom.xml
@@ -35,7 +35,7 @@
 		<!-- ################################ testing dependencies ################################ -->
 		<dependency>
 			<groupId>org.eclipse.sw360.antenna</groupId>
-			<artifactId>core-common-testing</artifactId>
+			<artifactId>model-testing</artifactId>
 			<version>${project.version}</version>
 			<scope>test</scope>
 		</dependency>
