+++ antenna/frontend-stubs-testing/src/main/java/org/eclipse/sw360/antenna/frontend/testProjects/P2TestProject.java
