+++ antenna/model-testing/src/main/java/org/eclipse/sw360/antenna/testing/AntennaTestWithMockedContext.java
