+++ antenna/model-testing/src/main/java/org/eclipse/sw360/antenna/testing/TestLicenseKnowledgeBase.java
@@ -9,7 +9,7 @@
  * SPDX-License-Identifier: EPL-2.0
  */
 
-package org.eclipse.sw360.antenna.knowledgebase;
+package org.eclipse.sw360.antenna.testing;
 
 import org.eclipse.sw360.antenna.api.ILicenseManagementKnowledgeBase;
 import org.eclipse.sw360.antenna.api.IProcessingReporter;
