+++ modules/p2/antenna-manifest-resolver/pom.xml
@@ -46,7 +46,7 @@
         <!-- ################################ testing dependencies ################################ -->
         <dependency>
             <groupId>org.eclipse.sw360.antenna</groupId>
-            <artifactId>core-common-testing</artifactId>
+            <artifactId>model-testing</artifactId>
             <version>${project.version}</version>
             <scope>test</scope>
         </dependency>
