+++ pom.xml
@@ -45,9 +45,11 @@
 
     <modules>
         <module>antenna/model</module>
+        <module>antenna/model-testing</module>
         <module>antenna/core</module>
         <module>antenna/core-workflow-steps</module>
         <module>antenna/frontend-stubs</module>
+        <module>antenna/frontend-stubs-testing</module>
         <module>antenna/abstract-compliance-checker</module>
         <module>modules/maven</module>
         <module>modules/sw360</module>
